requirements-dev.txt
-r requirements.txt
nose
doctest-ignore-unicode
pyyaml